+++ Tail/Player.h
@@ -10,7 +10,7 @@ public:
     Player();
     ~Player();
     virtual void Update();
-    virtual void FixedUpdate(sf::RenderWindow& window, double delta, const Level& level);
+    virtual void FixedUpdate(sf::RenderWindow& window, float delta, const Level& level);
 };
 
 #endif
